Removed useefect from color picker
